Allow object removal to be marked as temporary.
Add an optional flag to removeObject() to allow the removal to be
marked as temporary. If the flag is true, a removal event is still
generated, but it is marked as transient instead of permanent.

This is useful if you want to remove and reinsert an object or if you
manage a set of query handlers and know the object is going to be
inserted somewhere else such that the removal really will be
transient.