Add temporary flag to LocationUpdateListener::locationDisconnected.
Add a flag that allows the LocationServiceCache to indicate that the
removal it is notifying the listener about was marked as
temporary. This allows object removals to be marked as transient if
their removal from the object cache was marked as temporary. This is
useful when your location service cache is itself populated by a
query, which can result in quick removal/addition pairs.