Add node iterators.
Add nodesBegin() and nodesEnd() methods to QueryHandlerBase. Nodes are
objects or collections of objects that are created by the query
handler and may be exposed to the user. For some query handlers this
ends up just being the set of objects (e.g. brute force) and for others it
will be the nodes in the tree (e.g. RTree). They aren't guaranteed to
have a particular structure: they could be a set, a tree, or a forest
(e.g. rebuilding query handler sometimes has two active trees).

The implementations are straightforward but have a wrapper to make the
iterators appear like value types. To avoid lots of copying of the
internal iterator (since each implementation has to iterate over
different data structures), the wrapper performs copy-on-write when
there are more than two wrappers referencing the underlying
iterator. This should get rid of most of the allocations and make
these iterators appear like normal STL iterators.

To test these iterators add a key handler for 'n' which prints
nodes. Currently the iterators only expose the object or aggregate ID,
so this only prints those values.