Rework build system to support parent projects.
Rework the build system to support calling ADD_DIRECTORY() on it to
use it directly as part of the build of a parent project. Use three
optional settings from the parent project to control whether to build
the proxsim binaries, the type of RTree data to maintain (bounds or
max size), and whether to turn on validation for debugging.

Make the build able to reuse boost settings from the parent project,
or find its own. Only perform searches for OpenGL libraries if the
proxsim binaries were requested. Finally, export include directory and
additional definitions as directory properties so parent projects can
extract and use them.