Add compile-time option to control use of cut lifting.
Add an option to the build system to toggle between cut lifting and
more aggressive cut-maintainence. Cut lifting is the existing approach
where cuts are entirely removed from the subtree being modified so
they won't interfere with the operation. This is inefficient since it
causes cuts to have to be re-refined, generates unnecessary result set
changes, and looks bad to clients since they can get a large stream of
unnecessary updates and have to aggressively mask the changes.

The alternative is harder to implement but behaves more nicely: only
move cuts as required by the operation. The option toggles this new
mode. This patch only swaps out the implementation for tree
condensation, where nodes are possibly merged after a removal
occurs. This new approach uses the old lifting utilities, but only
moves cuts that pass through deleted nodes up to the parent. If they
only pass through descendents, they aren't lifted and the existing cut
is guaranteed to be valid by maintaining the ordering of grandchildren
as they are moved into their new parent.