Add support for insertion/splitting nodes without cut lifting.
Add a new version of insertion/splitting nodes that doesn't need cut
lifting. This is used with the recently added RTREE_LIFT_CUTS flag set
to false.

The majority of the changes are in RTreeCore, replacing the splitting
and tree adjustment algorithms, and in Cut.hpp, which handles updating
the cut and results after the split. Other changes are relatively
minor, mainly taking care of adding support for the additional
RTree callback required to make this work. A few other modifications,
like the ability to choose where in a node to insert a child are also
added.

The approach to splitting up children remains the same: collect the
children, choose two seeds, and then assign each object into one of
the node using the seeds. The changes occur in how we then cleanup the
rest of the tree and handle cascading splits. The basic approach is to
work our way up the tree, extending cuts as we split new nodes. This
gets each level where a cut passes through to have a full cut across
the tree. However, we also need to deal with rearranging cuts since
splitting a node high in the tree causes it's children, and therefore
all cuts passing through any descendents, to become jumbled. This is
resolved by finding contiguous segments of cuts below each child,
getting them listed in the order we want them, and then delegating the
task of reordering them to the cut. By performing this at each step up
the tree, we maintain cuts that properly cross the entire tree and are
in the correct order, even when the higher levels of the tree aren't
fully connected to each other yet (since there are still outstanding
split nodes).

Besides handling reordering itself to match the new layout of the
tree, the Cut class is also updated to handle changes to the result
set. Since cuts aren't lifted, the results will change differently
than before. We handle this as part of handling a split. Besides
adding a new cut node, we also check how to change the result set,
either adding just the new node (if the original node was in the
result set, then the new node is also valid as a result) or adding all
the children (if the original was refined, both can start
refined). This still isn't as efficient as possible because the split
code in RTreeCore is actually removing results from the cut when the
splitting of objects between the nodes is happening because it
triggers callbacks. If we can get rid of this then we'd only have one
object in one event. Currently we have something that looks more like
a lift up one level, then refinement back down again.