Fix missing removal notifications and order of removal/additions.
This fixes a couple of issues where aggregate notification of child
additions/removals could be incorrect. Generally within a single tree,
an addition should only occur after a removal (though ideally
immediately after the removal).

In condense_tree, we were re-inserting objects and nodes before
removing them from their parents, which avoided having to copy state
out, do the removal, then do the addition. This changes that code to
copy the data out first to ensure the correct removal -> addition
ordering. In one version, it's not currently possible to encounter an
issue with this as nodes are guaranteed to be empty. For that case,
this just leaves a pair of warnings, one where the additions/removals
occur and one in the earlier loop that decides what to remove which
contains the condition which guarantees that piece of code cannot
currently cause this misordering to occur.

In RTreeNode::clear, we were only notifying when objects were removed
from their parents. We just need to also loop through all child nodes
and notify aggregate listeners of their removal as well before
resetting the count to 0.