Add initial support for reconstructing trees.
Add methods for adding nodes and objects with specified parents,
starting with an empty tree, callbacks to update queries when a root
is created, and expose all of this via methods on the query handlers.

Also includes a fix to the rebuilding query handler node iterator to
make it work if it starts out initialized with invalid iterators
(e.g. primary begin == primary end because there are no nodes in
primary).