Make updates work for tree replication.
We want to support two distinct modes for queries: query results (a
cut or leaf nodes as the active set) and tree replication (all nodes
up to the cut as the active set). This adds support for that by
filtering the events that are sent so that, e.g., when doing tree
replication we don't send removal events when we refine a
node. Currently the choice of which to use is implicit, based on the
query handler, since tree replication only applies to manual queries
currently.