Make cascading splits work top down.
When an object is inserted, it can cause a cascade of
splits. Previously we just worked our way up the tree, splitting nodes
as necessary and cleaning up the tree and cuts as we went. Now that we
want to replicate trees, we need to make sure the sequence of events
is produced in a way that will be usable on the replicating node
without having to deal with invalid intermediate states. Since the old
approach worked bottom up, we would end up with disconnected
components as nodes were split but not yet connected back into the
tree.

This changes the approach by scanning up the tree determining all the
splits that will need to occur before performing any. It then works
its way top down, making each split add a new node, then rearrange the
children, and then finally insert the final object that caused the
overflow. This results in operations in an order which can be
reproduced on the client without ever leaving the tree in an invalid
intermediate state.

This should be working for both cut-lifting and non-cut-lifting
versions. Making it work for both required splitting the split_node
operation into 3 distinct parts -- a setup phase (for lifting cuts),
the main stage (generating the new node, attaching it, and rearranging
children) and a cleanup (notifying of the split and taking care of
fixing up cuts).