Don't try to refine empty nodes.
In replicated trees, you can hit nodes which satisfy the query but
have no children. Because we were previously just checking for leaves
and valid trees always had child nodes or objects, empty nodes weren't
a problem. With replicated trees, you could lose parts of the cut (or
the whole thing) by refining a cut at a node which had no
children. This just avoids doing that, moving the cut processing
forward if necessary. Also adds a few assertions to make sure we don't
try to refine nodes when it wouldn't be safe.