Add notification to LocationServiceCache of new imposters.
Add a method to LocationServiceCache to notify it when a new imposter
is created, making sure the cache is aware of it before any other
operations with it occur. This is tied to notification of aggregates
because that's the only way an external user could become aware of the
imposter. This change is necessary because the
AggregateListener::aggregateCreated call is made immediately and other
objects may want to use the LocationServiceCache to reliably get
information about the object, but since they may have to wait for it
to filter through the LocationService (possibly on a separate thread)
and then into the LocationServiceCache, they may fail to get the
information they want. This just ensures it is placed in the
LocationServiceCache and usable immediately, then superceded when it
makes it through the original way.