Ensure objects are tracked in the LocationServiceCache while they are in QueryEvents.
Make QueryEvents track the LocationServiceCache their parents come
from and make sure the LocationServiceCache maintains a copy of their
data until the QueryEvent is destroyed. This ensures that results
always have valid data to be used. Without this it was possible, for
example, to get a set of query events for which some aggregates had
been created, reported to the querier, removed, and destroyed. By the
time the querier is looking at the results some elements were missing.

This also requires additional methods in LocationServiceCache that
expose a raw reference counting interface without requiring the
Iterator resulting from regular tracking to be stored and used to
reduce the tracking refcount.