Fix ObjectLocationServiceCache to handle refcounting.
It was asserting that an object did not already exist, but the use of
refcounting means that it's possible the object will be held in the
cache after its removal was requested. Instead, allow udpating but
assert that it is marked as being dead and only being cached due to
non-zero refcount.