Fix a missing addition and removal for replicated trees with cuts through the node.
When replicated trees had nodes split, cuts were being updated to add
a new node, but the events weren't properly being added to the
results, which are needed so the children that are put into the new
node will have a parent.

This is progress towards getting replicated trees to work properly
without lift cutting (removing the prepare step where we still lift
cuts for the sake of replicated trees), but still doesn't fix it. At a
minimum we'll still need to fix issues with moving entire trees around
when splits occur above cuts, without having to delete and recreate them.