Add cxxtest and incorporate, starting with QueryCacheTest.
Add a submodule for cxxtest, add build support for finding and using
cxxtest, add a libprox_tests binary. Add a simple set of tests for
QueryCache to start with, which revealed a minor typo bug in
QueryCache::remove since it isn't currently used in the implementation
so it was never instantiated and compiled without the test.