Add simple tests for regular geometric (solid angle) query handlers.
Add TestLocationServiceCache to provide a LocationServiceCache for
tests. It isn't thread-safe and runs notifications immediately. Add
GeomHandlerTest as base class for BruteForce, RTree, RTreeCut, and
RTreeCutAgg HandlerTests since they all perform the same
tasks. However, it has to be careful about testing expected result
counts since RTreeCutQueryHandler with aggregates enabled can return
more results than the handlers which are answering the query
precisely. This doesn't thoroughly test the correctness of queries
across a variety of queriers and object sizes, just that the basic
functionality of getting (and removing) results works.