Add test for removal of nodes along cut, fix when not lifting cuts.
Add a test that checks correctness of replication when nodes along the
cut are removed. This exposed a problem where we were lifting cuts
when it wasn't necessary. This fixes that so the node is simply
removed via another cut callback. However, for simplicity, if removing
the cut node would require adding the parent (because the node is the
last child and not doing so would leave a gap in the cut), we fall
back to cut lifting in order to keep the new callback simpler. This
also changes another location we were still using cut lifting
unneccessary, in RTree_condense_tree, to use the same approach.