Add test and fix explicit addition of nodes along the cut.
Adding nodes directly (i.e. in a replicated tree) when there are cuts
running through siblings or siblings' ancestors had not been
updating cuts in anyway. This adds a test of that functionality,
making sure we just add a new cut node (lifting should never be
necessary).  A new callback is needed to pass this information up to
the cut, but the implementation is straigtforward as we can guarantee
the closest cut node is the one this is invoked on.