Re-choose insertion node when chosen leaf requires split.
Previously the code just continued on with insertion after splitting a
leaf node, but it actually needs to choose the insertion point again
because either of the two split nodes could be better. We need to make
sure to choose between only the two split nodes because if we just
reran selection, we could end up choosing some other node and causing
additional splits.