Fix behavior of object insertions at nodes with manual queries.
Add a test of object insertion in nodes with manual query cuts passing
through them. This revealed that inserting an object would refine the
cut incorrectly. Fix this by changing the satisfiesQuery() method of
the cut to always return false since it is only used in this one
context and we want it to behave as if the query was not satisfied.