Apply location updates to replicated nodes.
We need to apply updates to replicated nodes or trees that aren't
fully replicated wouldn't have the right information as they would
never be updated. To keep things simple and consistent, we only apply
updates to nodes that are currently leaves (locally, not in the
original tree), and then force recomputation up the rest of the tree.