Pass along reparent events to cuts for replicated queries.
The cut event handlers for reparent events only passed along
information about additions and removals to adjust regular cuts when a
reparent event was handled from the location service cache. However,
when a replicated query handler is used on top of a location service
cache generated from an original replicated query handler, reparenting
is important to the final replicated copy as well. This adds the
necessary events when the reparent events are desired (i.e. manual
query handlers).