Add LevelQueryHandler.
Add LevelQueryHandler, a cut-based query handler that tries to keep
the cut at a given level of the tree, e.g. an level cut at the root
(level 0), root's children (level 1), etc. Currently it abuses the
solid angle query parameter to specify the depth since we don't have a
generic interface for query parameters. Includes a basic test of the
functionality for a small tree.