Make RTree node data extensible by users.
Refactor the templates for different types of node data to not depend
on the CutNode type in the class type (only for some methods called
directly by RTree methods). Use this to make RTree-based query
handlers take an additional template parameter specifying the per-node
data class to use, allowing the user to customize the type of data and
some implementation details of the RTree (e.g. selection of insertion
nodes). Also update the tests and proxsim binaries to provide this
setting. The CMake setting still remains, but is only used to
customize the tests and proxsim binaries.