Don't provide default types for query handler implementations.
Providing defaults is now somewhat error prone because the
SimulationTraits come first, which had a reasonable default value, but
now we also have an RTree per-node data type for many of them. To
maintain the default SimulationTraits we also have to specify a
default for the per-node data. Because of the transition (and even if
someone just forgets to specify the per-node data type), it's too easy
to end up accidentally using the wrong type (e.g. just bounding sphere
instead of bounding sphere with maximum object size) by omitting the
template parameter.