RebuildingQueryHandler doesn't need a NodeDataType parameter.
This got changed with the other query handlers, but like the
QueryHandler base classes and BruteForceQueryHandler, the parameter
doesn't make sense and is unnecessary for this class since it doesn't
necessarily have a child QueryHandler with NodeData. Since it
delegates creation of new QueryHandlers, it doesn't matter if a child
QueryHandler does need the template parameter.