Fix edge case in cut reparent event handling.
Fix an edge case where we could get stuck in an infinite loop with
infinitely expanding cut if the portion of a cut that needed to be
moved ended on the same element as the location it needed to be moved
to. The fix is just to ignore this case because it should be a nop.