Pass leaf object events through AggregateListener.
Add aggregateObjectCreated and aggregateObjectDestroyed events to
AggregateListener to indicate when individual leaf objects are created
and destroyed in the tree. This lets you properly reconstruct the
entire tree instead of relying on implicit information about leaf
objects. The naming is a bit unfortunate -- the prefix aggregate is to
scope the methods for AggregateListener and ObjectCreated/Destroyed
indicates we're talking about leaf objects.