Allow user to force synchronous rebuilds.
For simple experiments run with the simulator, synchronous rebuild is
necessary since having queriers split across multiple query handlers
and reporting up to 2x the total ticks because both query handlers
print it out screws up the parsing of the output. Add a method for the
rebuilding query handler that lets the caller for the thread to wait
until the rebuild finishes and they've transitioned the queries so no
tick() calls see the intermediate states.