Track whether cut nodes are in the result set.
Use a bit in the CutNode to track whether it's currently in the result
set. This makes checking membership for those nodes faster. However,
this is only a modest improvement for now because the leaf objects
still require lookups because we don't have enough space anywhere to
save whether they are in the results yet.