Use remove/tryRemove on cuts depending on situation.
Distinguish between remove, which should definitely have the object to
be removed in the result set, and tryRemove, which tests if it is
there, removes it if it is, and returns whether it was removed. This
fixes a crashing assert since in some cases we were asserting that the
object was in the results when we were only trying to remove it if it
was.